+++ common/src/python/uploads/provenance.py
@@ -15,9 +15,10 @@ class FileProvenance(BaseModel):
     file_id: str
     file_name: str
     flywheel_path: str
+    created_date: str
     modified_date: str
 
-    @field_validator("modified_date")
+    @field_validator("created_date", "modified_date")
     def validate_iso_format(cls, v: str) -> str:
         """Ensure the modified date is in ISO format."""
         try:
@@ -35,10 +36,22 @@ class FileProvenance(BaseModel):
         cls, proxy: FlywheelProxy, parent: FileEntry
     ) -> "FileProvenance":
         """Create file provenance from parent FileEntry."""
+        # get the first version of the file to get the true creation
+        # date. this is done because the created date of the current
+        # file only corresponds to that version, which may be a reupload
+        versions = proxy.get_file_versions(parent.file_id)
+        if not len(versions) > 1:
+            raise ValueError("parent file has no file version history")
+
+        first_version = versions[0]
+        if first_version.version != 1:
+            raise ValueError("Version 1 of parent file not in expected position")
+
         return FileProvenance(
             file_id=parent.file_id,
             file_name=parent.name,
             flywheel_path=proxy.get_lookup_path(parent),
+            created_date=first_version.created.isoformat(),
             modified_date=parent.modified.isoformat(),
         )