+++ docs/csv_subject_splitter/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this gear are documented in this file.
 
 ## 1.0.7
 
-* Adds the parent's modified date to the `provenance` metadata
+* Adds the parent's first creation and last modified date to the `provenance` metadata
 * Updates to support all types of delimiters, not just empty spaces
 * Updates to Python 3.12 and switches to use `fw-gear` instead of `flywheel-gear-toolkit` (now deprecated)